refactor: Update DTO address, fix brand creation, adjust methods and fix endpoint inconsistencies

+++ backend/src/main/java/com/example/demo/repository/IAddressRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface IAddressRepository extends JpaRepository<Address, Integer> {
+public interface IAddressRepository extends JpaRepository<Address, Long> {
 
 
     List<Address> findAllByUserFk(Long userFk);
