+++ backend/src/main/java/com/example/demo/repository/IOrderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface IOrderRepository extends JpaRepository<Order, Integer> {
-    List<Order> findByUser_id(int id);
+public interface IOrderRepository extends JpaRepository<Order,Long> {
+    List<Order> findByUser_id(Long id);
 }
