+++ backend/src/main/java/com/example/demo/repository/IOrderStatusRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.entity.OrderStatus;
 import com.example.demo.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IOrderStatusRepository  extends JpaRepository<OrderStatus, Integer> {
+public interface IOrderStatusRepository  extends JpaRepository<OrderStatus, Long> {
 }
