+++ backend/src/main/java/com/example/demo/repository/IShippingMethodRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.entity.ShippingMethod;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IShippingMethodRepository  extends JpaRepository<ShippingMethod, Integer> {
+public interface IShippingMethodRepository  extends JpaRepository<ShippingMethod,Long> {
 }
